+++ src/SMAPI.Toolkit/Framework/ModScanning/ModFolder.cs
@@ -25,10 +25,10 @@ namespace StardewModdingAPI.Toolkit.Framework.ModScanning
         public Manifest Manifest { get; }
 
         /// <summary>The error which occurred parsing the manifest, if any.</summary>
-        public string ManifestParseError { get; }
+        public ModParseError ManifestParseError { get; set; }
 
-        /// <summary>Whether the mod should be loaded by default. This is <c>false</c> if it was found within a folder whose name starts with a dot.</summary>
-        public bool ShouldBeLoaded { get; }
+        /// <summary>A human-readable message for the <see cref="ManifestParseError"/>, if any.</summary>
+        public string ManifestParseErrorText { get; set; }
 
 
         /*********
@@ -39,16 +39,24 @@ namespace StardewModdingAPI.Toolkit.Framework.ModScanning
         /// <param name="directory">The folder containing the mod's manifest.json.</param>
         /// <param name="type">The mod type.</param>
         /// <param name="manifest">The mod manifest.</param>
+        public ModFolder(DirectoryInfo root, DirectoryInfo directory, ModType type, Manifest manifest)
+            : this(root, directory, type, manifest, ModParseError.None, null) { }
+
+        /// <summary>Construct an instance.</summary>
+        /// <param name="root">The root folder containing mods.</param>
+        /// <param name="directory">The folder containing the mod's manifest.json.</param>
+        /// <param name="type">The mod type.</param>
+        /// <param name="manifest">The mod manifest.</param>
         /// <param name="manifestParseError">The error which occurred parsing the manifest, if any.</param>
-        /// <param name="shouldBeLoaded">Whether the mod should be loaded by default. This should be <c>false</c> if it was found within a folder whose name starts with a dot.</param>
-        public ModFolder(DirectoryInfo root, DirectoryInfo directory, ModType type, Manifest manifest, string manifestParseError = null, bool shouldBeLoaded = true)
+        /// <param name="manifestParseErrorText">A human-readable message for the <paramref name="manifestParseError"/>, if any.</param>
+        public ModFolder(DirectoryInfo root, DirectoryInfo directory, ModType type, Manifest manifest, ModParseError manifestParseError, string manifestParseErrorText)
         {
             // save info
             this.Directory = directory;
             this.Type = type;
             this.Manifest = manifest;
             this.ManifestParseError = manifestParseError;
-            this.ShouldBeLoaded = shouldBeLoaded;
+            this.ManifestParseErrorText = manifestParseErrorText;
 
             // set display name
             this.DisplayName = manifest?.Name;